requirements.txt
Django==4.2
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
redis==4.6.0
asgiref==3.6.0
coverage==7.2.7
Pillow==9.5.0
sqlparse==0.4.4
requests==2.31.0
